More work on the store() method
Currently has an array of allowed fields, gets the data from the post and then makes a new array based on the allowed fields, user_id is hardcoded but will be changed eventually, and then we run array_map to run a function on all values of the array. Two inspects at the bottom to check results.