Refactor to allow middleware
Since we now have a login, some pages are locked to guests or authorized users. Added  param to registerroute, saves it on the routes, each HTTP function inside also takes in  as a param. Final change was on the route method, as we loop through the middleware (because  is an array) and then handle it in the authorize class. If the user's current role matches with the needed roles, all is well,otherwise redirect to the login page.